implement health bar
The health bar is implemented as a spritesheet.  In theory, it would be faster to implement as an image, but as far as I could tell in the phaser docs, there is no way to dynamically crop an image on the screen.  Therefore, I had to put 5 sprites on the screen and fix them to the camera by setting the property fixedToCamera.  The update function then makes each sprite invisible as needed by player damage.

Hopefully I will think of a better way to implement this, because this solution is a bit clunky IMO.